got transformer all good now

+++ src/main/java/com/method/donuts/bos/method/payments/Payment.java
@@ -14,7 +14,7 @@ public class Payment {
     public Object destination_trace_id;
     public String source;
     public String destination;
-    public int amount;
+    public Float amount;
     public String description;
     public String status;
     public Object error;
